Short-circuiting and add `conflicting_inputs.rs`
