+++ src/lib.rs
@@ -16,6 +16,7 @@ pub mod axislike;
 pub mod buttonlike;
 pub mod clashing_inputs;
 pub mod common_conditions;
+pub mod conflicting_inputs;
 mod display_impl;
 pub mod errors;
 pub mod input_map;
@@ -45,9 +46,9 @@ pub mod prelude {
     pub use crate::input_mocking::{MockInput, QueryInput};
     pub use crate::user_input::{InputKind, Modifier, UserInput};
 
+    pub use crate::conflicting_inputs::TrackingInputType;
     pub use crate::plugin::ToggleActions;
     pub use crate::plugin::{InputManagerPlugin, InputManagerSystem};
-    pub use crate::systems::TrackingInputType;
     pub use crate::{Actionlike, InputManagerBundle};
 }
